fix: activity loading and rename file
